[serve] Drop handle metrics from dead actors
Drop metrics from handles that lived on dead serve actors. This doesn't cover everything (e.g. handles that live in driver scripts, or tasks, or any non-serve actors), but it should cover most cases, and the remaining cases will be handled by the default 20s timeout invalidation.

Handles living in the proxies and handles living in Serve replicas are identified as the handles living on Serve actors. At any point if the actor on which these handles reside is not alive anymore, the controller will drop metrics from those handles.

Signed-off-by: Cindy Zhang <cindyzyx9@gmail.com>