[serve] prevent growing memory in replica autoscaling metrics
In the https://github.com/ray-project/ray/pull/44638 refactor, the code that prevents the dict of replica autoscaling metrics living in the controller from growing in memory over time (as replicas upscale/downscale) was missed and didn't get ported over. This PR patches that.

I also added to the unit test `test_basic_autoscaling` to make sure dead replicas are removed from the in-memory dict of replica metrics.


Signed-off-by: Cindy Zhang <cindyzyx9@gmail.com>