[runtime_env] change container manager to unified plugin mechanism
Currently, we create a custom `ContainerManager` class and invoke it specially with `container_manager.setup()`. This doesn't conform with the runtime env plugin pattern we use for every other runtime env field.

This PR removes `ContainerManager` and moves the container code to `ContainerPlugin` which extends the base `RuntimeEnvPlugin` class. It changes the callsite to run `create_plugin_if_needed` instead of `container_manager.setup()`. This way, in the future if we add support for using `container` together with other runtime env fields, the API is more unified.

Signed-off-by: Cindy Zhang <cindyzyx9@gmail.com>