[runtime-env] automatically infer worker path when starting worker in container
Instead of requiring user provide the path to `default_worker.py` in the container, which is bad UX, we should be able to automatically detect where it is. During runtime environment setup, pull the image ahead of time and run a python command to import `ray._private.workers.default_worker` and get its file path.


Signed-off-by: Cindy Zhang <cindyzyx9@gmail.com>